backend/requirements.txt
fastapi
uvicorn[standard]

# For Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0